requirements.txt
numpy == 1.16.4
scipy == 1.3.0
scikit_learn == 0.17.1
bert-serving-client == 1.10.0                
bert-serving-server == 1.10.0  
pysparnn  == 0.4  
Keras  == 2.2.4                 
Keras-Applications  == 1.0.8                 
Keras-Preprocessing  == 1.1.0  